consolidate arbitrary impls to one file in zebra-chain
